Move <Drawing/> inside <Drawings/>
