+++ src/Dashboard/Drawings/Drawings.tsx
@@ -2,9 +2,15 @@ import React, { FC } from 'react';
 import { Link } from 'react-router-dom';
 
 import Button from 'shared/components/Button/Button';
+import Icon from 'shared/components/Icon/Icon';
 
-import Drawing from '../Drawing/Drawing';
-import Wrapper from './Styles';
+import {
+  Wrapper,
+  Drawing,
+  StyledLink,
+  Details,
+  Title,
+} from './Styles';
 
 const Drawings:FC = () => (
   <Wrapper>
@@ -13,7 +19,20 @@ const Drawings:FC = () => (
     </Link>
 
     <div>
-      <Drawing />
+      <Drawing>
+        <StyledLink to="/editor">
+          <Details>
+            <Icon variant="file" size="1em" />
+            <Title>Drawing #5</Title>
+          </Details>
+        </StyledLink>
+
+        <Button
+          disabled
+          icon="moreHorizontal"
+          variant="empty"
+        />
+      </Drawing>
     </div>
   </Wrapper>
 );
